requirement.txt
asn1crypto==1.5.1
cryptography==41.0.7
pyOpenSSL==23.3.0